+++ resources/views/examples/color-picker/default.blade.php
@@ -1,0 +1,1 @@
+<x-ui.color-picker name="brand_color" value="#6366f1" />