+++ resources/views/examples/color-picker/inline.blade.php
@@ -1,0 +1,1 @@
+<x-ui.color-picker value="#22c55e" inline />