+++ resources/views/examples/editable/default.blade.php
@@ -1,0 +1,1 @@
+<x-ui.editable name="title" value="Project Apollo" label="title" class="max-w-xs" />