+++ resources/views/examples/editable/with-placeholder.blade.php
@@ -1,0 +1,1 @@
+<x-ui.editable name="nickname" label="nickname" placeholder="Add a nickname…" class="max-w-xs" />