+++ resources/views/examples/file-upload/default.blade.php
@@ -1,0 +1,3 @@
+<div class="w-full max-w-md">
+    <x-ui.file-upload name="document" maxSizeLabel="Up to 10MB" />
+</div>