+++ resources/views/examples/file-upload/image-preview.blade.php
@@ -1,0 +1,3 @@
+<div class="w-full max-w-md">
+    <x-ui.file-upload name="photos" multiple accept="image/*" maxSizeLabel="PNG, JPG or GIF" />
+</div>