+++ resources/views/examples/file-upload/multiple.blade.php
@@ -1,0 +1,3 @@
+<div class="w-full max-w-md">
+    <x-ui.file-upload name="attachments" multiple accept=".pdf,.doc,.docx,.txt" maxSizeLabel="Max 5 files" />
+</div>