+++ resources/views/examples/number-input/default.blade.php
@@ -1,0 +1,1 @@
+<x-ui.number-input name="quantity" :value="1" class="w-40" />