+++ resources/views/examples/number-input/disabled.blade.php
@@ -1,0 +1,1 @@
+<x-ui.number-input name="quantity" :value="3" :disabled="true" class="w-40" />