+++ resources/views/examples/number-input/min-max.blade.php
@@ -1,0 +1,1 @@
+<x-ui.number-input name="guests" :value="2" :min="1" :max="10" class="w-40" />