+++ resources/views/examples/number-input/sizes.blade.php
@@ -1,0 +1,5 @@
+<div class="flex flex-col items-start gap-4">
+    <x-ui.number-input name="qty-sm" size="sm" :value="1" class="w-36" />
+    <x-ui.number-input name="qty-default" size="default" :value="1" class="w-40" />
+    <x-ui.number-input name="qty-lg" size="lg" :value="1" class="w-44" />
+</div>