+++ resources/views/examples/number-input/step.blade.php
@@ -1,0 +1,1 @@
+<x-ui.number-input name="rate" :value="2.5" :step="0.5" :min="0" class="w-40" />