+++ resources/views/examples/password-strength/default.blade.php
@@ -1,0 +1,3 @@
+<div class="w-full max-w-sm">
+    <x-ui.password-strength name="password" label="Password" />
+</div>