+++ resources/views/examples/password-strength/with-checklist.blade.php
@@ -1,0 +1,3 @@
+<div class="w-full max-w-sm">
+    <x-ui.password-strength name="new_password" label="New password" :min-length="10" :show-checklist="true" />
+</div>