+++ resources/views/examples/tags-input/default.blade.php
@@ -1,0 +1,3 @@
+<div class="w-full max-w-sm">
+    <x-ui.tags-input name="tags" />
+</div>