+++ resources/views/examples/tags-input/max-tags.blade.php
@@ -1,0 +1,3 @@
+<div class="w-full max-w-sm">
+    <x-ui.tags-input name="topics" :value="['Design']" :max="3" placeholder="Up to 3 tags…" />
+</div>